% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iba_co1_data_fxn.R
\name{get_iba_co1_data}
\alias{get_iba_co1_data}
\title{Getting IBA CO1 data}
\usage{
get_iba_co1_data(
  country,
  dataset = "lysate|homogenate|soil|litter|ethanol",
  data_path = "~/dev/figshare-repos/iba/processed_data/v3/",
  metadata_path = "~/dev/figshare-repos/iba/raw_data/v6/",
  calibrate = FALSE,
  remove_spikes = TRUE,
  format = "data.frame"
)
}
\arguments{
\item{country}{The country for which data is requested (\code{"MG"} or \code{"SG"}).}

\item{dataset}{The dataset requested: \code{"lysate"}, \code{"homogenate"}, \code{"soil"},
\code{"litter"} or \code{"ethanol"}. See \emph{Details} for more information.}

\item{data_path}{The path to the processed data. Defaults to
\code{"~/dev/figshare-repos/iba/processed_data/v3/"}}

\item{metadata_path}{The path to the metadata (raw data). Defaults to
\code{"~/dev/figshare-repos/iba/raw_data/v6/"}}

\item{calibrate}{Whether to calibrate data using biological spike-ins.
Defaults to \code{FALSE}.}

\item{remove_spikes}{Whether to remove spike-ins. Defaults to \code{TRUE}.}

\item{format}{The desired data format ("data.table" or "data.frame")}
}
\value{
OTU read table in desired format with taxonomic annotation of OTUs
}
\description{
Retrieves the requested CO1 data from the IBA project, given paths to the
raw data and processed data. The retrieved data can be calibrated using
biological spike-ins, if desired, and spike-in data can either be removed
or kept, as desired.
}
\details{
The \code{dataset} parameter allows the user to select the desired data subset.
The available options are \code{lysate}, \code{homogenate}, \code{soil}, \code{litter} and
\code{ethanol}. For "MG" (Madagascar), only \code{lysate} and \code{litter} data are
available. It is possible to request several subsets at once by concatenating
the dataset names. For instance, \code{lysate|homogenate} would retrieve both
lysate and homogenate data.

Note that only biological spike-in data are used to
calibrate reads. Spike-ins are identified by occurrence patterns (occurring
in more than 80\% of samples) and taxonomic annotation (belonging to Insecta
if bioliogical spike-ins), calling the function \code{identify_spikes}.

The return value is an OTU read table that contains the full taxonomic
annotation of the OTU (using the recommended IBA annotation combining SINTAX
and EPA-NG annotations). The headers of the read columns are the
\code{sampleID_NGI} identifiers associated with the sequencing run in which the
data were generated. Thus, in a merged dataset it is not possible to
distinguish the data subsets without looking into the IBA metadata files.

Use the function \code{get_iba_sample_data} to get the associated field sample
and site metadata for those sequencing run identifiers. The function
\code{get_iba_taxon_data} can be used to retrieve such data together with read
numbers for specific taxa (individual OTUs or groups of OTUs with the same
taxonomic annotation at some rank).
}
\examples{
get_iba_co1_data(country="SE",dataset="lysate",calibrate=TRUE)
get_iba_co1_data(country="MG",dataset="lysate|homogenate")
}
