% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iba_co1_data_fxn.R
\name{handle_spikes}
\alias{handle_spikes}
\title{Handle spike-ins}
\usage{
handle_spikes(counts, samples, taxonomy, calibrate, remove_spikes)
}
\arguments{
\item{counts}{The OTU table as a data.frame or data.table object.}

\item{samples}{The names of the columns in the OTU table containing the
spike-in data, and for which the reads should (potentially) be calibrated.}

\item{taxonomy}{The taxonomic annotation of the OTUs.}

\item{calibrate}{Whether to calibrate data using biological spike-ins.}

\item{remove_spikes}{Whether to remove spike-ins.}
}
\value{
OTU read table in desired format with taxonomic annotation of OTUs
}
\description{
This function identifies biological and synthetic spike-ins. It then
calibrates the data using the biological spike-in data, if this is requested.
Finally, it removes the spike-ins, if this is requested.
}
\details{
In the OTU table passed in as \code{counts}, the first column is expected to
contain the OTU cluster names. The function handles both \code{data.frame} and
\code{data.table} formats for the OTU table. Note that data in \code{data.table} format
takes a long time to calibrate, at least on MacIntosh systems.

If some of the specified \code{samples} columns do not contain spike-in reads at
all, the read numbers are not changed.

The calibrated read numbers are returned as integers. The numbers are rounded
upwards to the nearest integer to make sure that presences are not turned
into absences for samples with few reads.
}
\examples{
handle_spikes(counts, samples, taxonomy, calibrate=TRUE, remove_spikes=TRUE)
}
