% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes_controls_fxns.R
\name{identify_control_clusters}
\alias{identify_control_clusters}
\title{Identify control_clusters}
\usage{
identify_control_clusters(counts, taxonomy, samples, controls, cutoff = 0.05)
}
\arguments{
\item{counts}{OTU table, with first column being cluster name.}

\item{taxonomy}{The taxonomic annotation (assumed to contain cluster and
taxonomic annotations for all ASVs)}

\item{samples}{Names (column headers) of the real samples.}

\item{controls}{Names (column headers) of the control samples.}

\item{cutoff}{Threshold for identifying control clusters, given in terms of the
proportion of control samples in which a cluster is allowed to occur before
it is removed. The default threshold is set to 0.05, meaning that clusters
occurring in more than 5% of control samples will be removed.}
}
\value{
A list containing a vector of identified control clusters and
a data frame containing detailed information on these clusters.
}
\description{
Identifies control clusters according to specified criteria.
}
