% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes_controls_fxns.R
\name{identify_spikes}
\alias{identify_spikes}
\title{Identify spike-ins}
\usage{
identify_spikes(counts, spikein_samples, taxonomy, cutoff = 0.8)
}
\arguments{
\item{counts}{OTU table with first column being OTU cluster name}

\item{spikein_samples}{The samples (column headers in the OTU table)
supposed to contain spike-in reads.}

\item{taxonomy}{The taxonomic annotations for the OTU clusters. If a complete
ASV taxonomy is provided, the annotation of the representative ASVs is used.}

\item{cutoff}{Cutoff for identifying spike-in clusters, given in terms of
the proportion of samples. The default cutoff is set to 0.8, meaning that
clusters occurring in more than 80% of `spikein_samples` are identified as
candidate spike-in clusters.}
}
\value{
A list giving all spikein clusters identified, as well as separate
vectors of biological and synthetic spikeins. The taxonomic annotation of
the identified biological spike-in clusters is also returned.
}
\description{
The function identifies biological and synthetic spike-ins using occurrence
pattern. Specifically, spike-ins are identified as clusters occurring in
more than the `cutoff` proportion of samples. For biological spike-ins, it
is also required that they are annotated as belonging to Insecta. Synthetic
spike-ins are identified as those candidate spike-in clusters missing a
taxonomic annotation.
}
