% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes_controls_fxns.R
\name{mean_max}
\alias{mean_max}
\title{Stats for OTU/ASV partitions}
\usage{
mean_max(counts, index)
}
\arguments{
\item{counts}{The OTU/ASV table as a data.frame or data.table object.}

\item{index}{Vector of booleans indicating which rows are considered for
removal.}
}
\value{
A list of statistics on the clusters to be removed or kept.
}
\description{
This function computes various statistics for the reads of a set of OTUs
or ASVs identified for potential removal (for instance, because they occur
in too many control samples). The set is specified as an index (a vector of
`TRUE` and `FALSE` values of the same length as the number of rows in the 
`counts`table, and should specify the rows identified for potential removal.
Statistics are provided for both the rows identified for removal and the rows
that would remain after removal.
}
