% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes_controls_fxns.R
\name{remove_control_clusters}
\alias{remove_control_clusters}
\title{Remove control_clusters}
\usage{
remove_control_clusters(
  filtered_counts,
  all_cluster_counts,
  taxonomy,
  samples,
  controls,
  cutoff = 0.05
)
}
\arguments{
\item{filtered_counts}{Filtered cluster counts; the first
column should be the cluster name.}

\item{all_cluster_counts}{Cluster counts for all samples (real samples and
control samples).}

\item{taxonomy}{The taxonomic annotation (assumed to contain cluster and
taxonomic annotations for all ASVs)}

\item{samples}{Names (column headers) of the real samples.}

\item{controls}{Names (column headers) of the control samples.}

\item{cutoff}{Threshold for removing control clusters, given in terms of the
proportion of control samples in which a cluster is allowed to occur before
it is removed. The default threshold is set to 0.05, meaning that clusters
occurring in more than 5\% of control samples will be removed.}
}
\value{
A list containing the OTU table with filtered counts, with control
clusters removed. The list also includes the taxonomy with control clusters
removed, as well as a data frame containing information on the removed
control clusters.
}
\description{
The function removes the control clusters according to specified criteria
}
