% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes_controls_fxns.R
\name{remove_spikes}
\alias{remove_spikes}
\title{Remove spike-ins}
\usage{
remove_spikes(counts, spikein_samples, taxonomy, cutoff = 0.8)
}
\arguments{
\item{counts}{OTU table with first column being OTU cluster name}

\item{spikein_samples}{The samples (column headers in the OTU table)
supposed to contain spike-in reads.}

\item{taxonomy}{The taxonomic annotations for the OTU clusters. If a complete
ASV taxonomy is provided, the annotation of the representative ASVs is used.}

\item{cutoff}{Cutoff for identifying spike-in clusters, given in terms of
the proportion of samples. The default cutoff is set to 0.8, meaning that
clusters occurring in more than 80\% of \code{spikein_samples} are identified as
candidate spike-in clusters.}
}
\value{
A list containing the counts table with spikeins removed, as well
as the taxonomic annotation of the biological spikeins identified.
}
\description{
Identifies the spikeins by calling the \code{identify_spikes} function, and then
removes them.
}
